refactor(ONB-204): rename nav component to menu and simplify density util
